Adding javascript language restrictions.
This is intended to block any usage of operators and keywords that have to do with either creating new objects or mutating them. This is supposed to deter users from adding complex logic into the templates. It is, of course, still possible for user to mutate the data objects with member methods.