Adding a "stack" for keeping arguments between open/close tags.
The arguments passed to the opening tags `{{ each(arr) "foo" }}` were lost when tag.suffix is rendered. This seems contrary to the intention of the original code, which tries to render arguments to suffixes without ever having any.

Extra points: Nesting is tested for tags which use both suffix & prefix.